split state
